Bug where passwords cleared in DB
There was a bug where if you wanted to be matched, or got matched and
logged out your password would be cleared in the database. I found the
bug on the logout branch but fixed it separately here. It occurs when
the passwords are null'd. For whatever reason when a user logs out they
get updated in the DB and since their password got assigned to null it
gets wiped. Now fixed.