Error msg on 0 picked, fixed list bug
When no places were selected, the match would continue with an empty
list. No longer happens now, we check for that. Also had to update the
click event to clear the error if its there. Found a bug in the list
where right clicking to remove would let you add the same restaurant
multiple time. Was just an invalid parse of the id to an int which is no
longer needed. Removing that fixed it.