Now [] restaurants count as match to any
Now if  a user has no restaurants in their list, they match to anyone.
Web client still doesn't allow < 1 choice (as this is ideal) but this
supports ios client now. On the server 2 people with no restaurants in
common will not match because there is no place to choose from. Also
changed progress bar to green.